aws-python-rest-api-with-dynamodb/requirements.txt
boto3==1.36.16
botocore==1.36.16
jmespath==1.0.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
s3transfer==0.11.2
six==1.17.0
urllib3==2.3.0
bcrypt==4.2.1